feat: a menu of the things you cannot already do with the hotkey
The menu bar had grown a row for everything the app owns, including two ways
into settings and a button for the one action the hotkey already does.

*Start Dictation* is gone. Dictation is a key you hold; a menu item for it is a
second entrance nobody walks through — you would have to let go of the mouse,
find the icon, click twice, and talk. It also had to be kept truthful, so
`updateUI` retitled it to *Stop Dictation* on every tick of a recording.

*Edit Config…* and *Settings & Permissions…* were two items for one question.
The window mostly described the config file the other item opened: which hotkey
is bound, in which mode, where recordings land — all of it read out of
config.yaml, none of it editable there, and free to drift from the file it was
paraphrasing. So *Settings…* opens config.yaml, which is where settings have
always lived, and the window keeps only what a file cannot answer: whether
macOS is letting this app hear you and type for you. `SettingsWindow.swift` is
`PermissionsWindow.swift`, two rows and nothing else.

*Permissions…* now appears only while a permission still needs doing —
microphone not granted, or Accessibility missing while `insert_mode: paste`
needs it to type. In clipboard mode nothing needs Accessibility, and an item
nagging about a permission your configuration never uses is noise. Checked as
the menu opens, because the answer changes in System Settings rather than here.

*About* is new: the version out of the bundle, stamped by release.sh, and a
link to the repository. With a dev build and a release build installed side by
side and a hotkey each, "which one is this and how old is it" had no answer
anywhere in the app — only in `defaults read` against the right Info.plist.

One system behaviour had to be turned off. macOS decorates menu items it
recognises, and puts a gear beside anything titled "Settings…", so the menu had
exactly one icon in it — which reads as a mistake rather than as emphasis. The
switch is `preferredImageVisibility`, and it arrives in the macOS 27 SDK while
this builds against 26, so it is set by name behind a `responds(to:)` check: a
no-op on any system without the property, applied to every item so nothing else
can sprout one either.

`lastTranscript` moved to AppDelegate on the way past. It is what a correction
or a spoken command works against when there is no selection — pipeline state
that happened to be parked in a view model, and the only reason that model was
reachable from half the file.

Not covered by anything: the menu is built inline and can only be checked by
clicking it, which is also why the gear needed a real machine to notice. Built,
installed and launched here; the item list and the About panel were read from
the source, not from a screenshot.
